fix: moves missing function.
